1) combine indian visitor Sujata's contribution, which let WRF-Var to use NMM firstguess. 2) Few bugs fix.
git-svn-id: https://svn-wrf-model.cgd.ucar.edu/trunk@3325 b0b5d27b-6f0f-0410-a2a3-cb1e977edc3d

+++ var/scripts/da_run_real.ksh
@@ -69,7 +69,12 @@ if $DUMMY; then
    done
    echo Dummy real > wrfbdy_d01
 else
-   ln -fs $REAL_INPUT_DIR/$DATE/met_em.d* .
+   if [[ -e $REAL_INPUT_DIR/$DATE ]]; then
+      ln -fs $REAL_INPUT_DIR/$DATE/met_em.d* .
+   else
+      LOCAL_DATE=$($BUILD_DIR/da_advance_time.exe ${DATE} -3 2>/dev/null)
+      ln -fs $REAL_INPUT_DIR/$LOCAL_DATE/met_em.d* .
+   fi
    ln -fs ${WRF_DIR}/main/real.exe .
    $RUN_CMD ./real.exe
    RC=$?
